image-recognition/requirement.txt
torch
transformers
pillow
matplotlib
numpy